pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pyperclip>=1.11.0",
    "python-chess>=1.999",
    "rich>=15.0.0",
]
